test(storage): KB hybrid search/pagination + row-level branch coverage (#437, #443) (#449)
Adds DB-gated integration coverage for the storage crate branches
flagged by the 2026-07 test-coverage audit, plus one production bug fix
the new tests surfaced.

Closes #437
Closes #443
Refs #445

## #437 — knowledge hybrid search (RRF + vector scoping), pagination,
delete_many

`crates/storage/tests/knowledge_hybrid_and_pagination.rs` (8 tests).
These feed **real embeddings** so `search`'s hybrid **vector branch
actually runs** — existing suites only ever hit the empty-embedding FTS
fallback and `search_text`, so the RRF fusion and the vector branch's
`WHERE user_id` scoping were never exercised.

- `knowledge_hybrid_search_is_user_scoped` — Bob searches with a
non-empty embedding; Alice's embedded doc must not leak through the
vector branch (`$6`).
- `knowledge_hybrid_search_excludes_tags` — `exclude_tags` (`NOT (tags
&& $7)`) drops a tagged doc reachable only via the vector branch.
- `knowledge_hybrid_search_rrf_orders_by_fused_rank` — a doc in both the
vector and text lists wins on fused RRF score.
- `list_page_walks_cursors_without_dup_or_gap`,
`list_page_tiebreaks_on_created_at_then_id` (both orders),
`list_page_rejects_malformed_cursor`, `list_page_clamps_limit_1_to_500`
— keyset pagination.
- `delete_many_ignores_foreign_ids` — cross-user opacity.

`crates/storage/tests/conversation_message_search.rs` (1 test):
- `search_messages_is_double_user_scoped` — message-FTS and
title/summary axes, both `user_id`-scoped, including a corrupted row
that only the `m.user_id` half hides.

## #443 — row-level error/lifecycle branches

`crates/storage/tests/storage_row_level.rs` (11 tests):
`create_duplicate_turn_id_errors`, `create_duplicate_task_id_errors`,
`create_summary_stamps_range_then_get_returns_it`,
`expand_summary_clears_summary_id`,
`archive_foreign_conversation_is_not_found`, `double_archive_is_ok`,
`tag_dedup_redirects_to_canonical`,
`resolve_active_name_follows_deprecation_without_cycling`,
`tool_registry_upsert_and_search`,
`unregister_source_removes_only_that_source`,
`migrate_json_imports_conversations_and_knowledge`.

## Bug fix (conversation.rs `archive`)

The `rows_affected == 0` existence probe decoded `SELECT 1` (Postgres
`INT4`) into a Rust `(i64,)`, which errored (`Rust type i64 … not
compatible with SQL type INT4`) **whenever a row actually matched** —
i.e. re-archiving your OWN already-archived conversation returned a
`Storage` error instead of `Ok`. The not-found path never decoded a row,
so the defect hid behind the happy-path tests. Decode into `(i32,)`.
`double_archive_is_ok` was RED against unmodified code and is now GREEN.

## Verification

- `TEST_DATABASE_URL=… cargo test -p desktop-assistant-storage` — 162
passed / 0 failed (20 new).
- `cargo fmt --check` clean; `cargo clippy -p desktop-assistant-storage
--all-targets` — 0 warnings.
- Every new correct-behavior test proven non-vacuous by a targeted
production mutation (all flipped RED, then reverted); the vector-branch
`$6` drop makes `knowledge_hybrid_search_is_user_scoped` RED as
required.

+++ crates/storage/src/conversation.rs
@@ -428,7 +428,14 @@ impl ConversationStore for PgConversationStore {
             // different user — `SELECT 1 …` distinguishes. The
             // existence probe is itself user-scoped so a cross-user
             // lookup still returns "not found" without leaking.
-            let exists: Option<(i64,)> =
+            //
+            // The literal `1` is Postgres `int4`, so it must decode into an
+            // `i32`: decoding into `i64` errored ("Rust type i64 … not
+            // compatible with SQL type INT4") only when a row actually
+            // matched — i.e. re-archiving your OWN already-archived
+            // conversation returned a Storage error instead of Ok. The
+            // not-found path never decoded a row, which is why the bug hid.
+            let exists: Option<(i32,)> =
                 sqlx::query_as("SELECT 1 FROM conversations WHERE user_id = $1 AND id = $2")
                     .bind(user_id.as_str())
                     .bind(&id.0)
